+++ com.unity.shadergraph/Tests/Editor/UnitTests/SerializedGraphTests.cs
@@ -647,7 +647,7 @@ namespace UnityEditor.Graphing.UnitTests
             graph.Connect(outputNode.GetSlotReference(TestableNode.Output0), inputNode.GetSlotReference(TestableNode.Input0));
             Assert.AreEqual(1, graph.edges.Count());
 
-            graph.RemoveElements(graph.GetNodes<AbstractMaterialNode>(), graph.edges, Enumerable.Empty<GroupData>());
+            graph.RemoveElements(graph.GetNodes<AbstractMaterialNode>().ToArray(), graph.edges.ToArray(), new GroupData[] {}, new StickyNoteData[] {});
             Assert.AreEqual(0, graph.GetNodes<AbstractMaterialNode>().Count());
             Assert.AreEqual(0, graph.edges.Count());
         }